requirements.txt
python-telegram-bot==13.7
urllib3==1.26.18
requests==2.32.5
environs==14.5.0

